% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cemitool.R
\name{cemitool}
\alias{cemitool}
\title{Full gene co-expression analysis}
\usage{
cemitool(exprs, annot, gmt, cor_method = c("pearson", "spearman"),
  merge_similar = TRUE, split_mods = FALSE, ora_pval = 0.05,
  min_ngen = 30, diss_thresh = 0.8, plot = FALSE, verbose = FALSE)
}
\arguments{
\item{exprs}{Gene expression \code{"data.frame"}.}

\item{annot}{Sample annotation \code{"data.frame"}.}

\item{gmt}{A character string with name of the Gene set file in GMT format.}

\item{cor_method}{A character string indicating which correlation coefficient is
to be computed. One of \code{"pearson"} or \code{"spearman"}.
Default \code{"pearson"}.}

\item{merge_similar}{Logical. If \code{TRUE}, merge similar modules.}

\item{ora_pval}{P-value for overrepresentation analysis. Default \code{0.05}.}

\item{min_ngen}{Minimal number of genes per submodule. Default \code{30}.}

\item{diss_thresh}{Module merging correlation threshold for eigengene similarity.
Default \code{0.8}.}

\item{plot}{Logical. If \code{TRUE}, plots all figures.}

\item{verbose}{Logical. If \code{TRUE}, reports analysis steps.}

\item{split_modules}{Logical. If \code{TRUE}, splits modules by correlation sign.}
}
\value{
just god knows
}
\description{
Defines co-expression modules and functionally characterizes
each one of them.
}
\examples{
cemitool(exprs=expression.df)

}

