% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\docType{methods}
\name{diagnostic_report}
\alias{diagnostic_report}
\alias{diagnostic_report,CEMiTool-method}
\title{Diagnostic report}
\usage{
diagnostic_report(cem, ...)

\S4method{diagnostic_report}{CEMiTool}(cem, title = "Diagnostics",
  directory = "./Reports/Diagnostics", force = FALSE, ...)
}
\arguments{
\item{cem}{Object of class \code{CEMiTool}.}

\item{...}{parameters to rmarkdown::render}

\item{title}{Character string with the title of the report.}

\item{directory}{Directory name for results.}

\item{force}{If the directory exists, execution will not stop.}
}
\value{
An HTML file with an interactive diagnostic report.
}
\description{
Creates report for identifying potential problems with data.
}
