% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\docType{methods}
\name{enrichment_report}
\alias{enrichment_report}
\alias{enrichment_report,CEMiTool-method}
\title{Enrichment analysis report}
\usage{
enrichment_report(cem_obj, ...)

\S4method{enrichment_report}{CEMiTool}(cem_obj, directory)
}
\arguments{
\item{cem_obj}{Object of class \code{CEMiTool}.}

\item{directory}{Directory name for results.}
}
\description{
Creates report for modules enrichment analysis results
}
\examples{

}

