% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cemitool.R
\docType{methods}
\name{expr_data}
\alias{expr_data}
\alias{expr_data,CEMiTool-method}
\alias{expr_data<-}
\alias{expr_data<-,CEMiTool-method}
\title{Retrieve and set expression attribute}
\usage{
expr_data(cem_obj)

\S4method{expr_data}{CEMiTool}(cem_obj)

expr_data(cem_obj) <- value

\S4method{expr_data}{CEMiTool}(cem_obj) <- value
}
\arguments{
\item{cem_obj}{Object of class \code{CEMiTool}}

\item{expr}{Object of class \code{data.frame} with gene
expression data}
}
\value{
Object of class \code{data.frame} with gene expression data
}
\description{
Retrieve and set expression attribute
}

