% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{expr_pct_filter}
\alias{expr_pct_filter}
\title{Filter genes based on expression.}
\usage{
expr_pct_filter(expr, pct = 0.75)
}
\arguments{
\item{expr}{expression file containing genes in the rows and samples in the columns}

\item{pct}{percentage of most expressed genes to maintain}
}
\value{
A data.frame containing the results
}
\description{
Filter genes based on expression.
}
\examples{
# Get example expression data
data(expr)
# Filter 80\% most expressed genes
filt_expr <- expr_pct_filter(expr, pct=0.8)
}
\keyword{internal}
