% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\docType{methods}
\name{filter_expr}
\alias{filter_expr}
\alias{filter_expr,CEMiTool-method}
\title{Filter gene expression table}
\usage{
filter_expr(cem_obj, ...)

\S4method{filter_expr}{CEMiTool}(cem_obj, pval = 0.05, n_genes)
}
\arguments{
\item{cem_obj}{Object of class \code{CEMiTool}}

\item{...}{other optional parameters}

\item{pval}{P-value cutoff for gene selection}

\item{n_genes}{number of genes to be selected}
}
\value{
Object of class \code{CEMiTool} with selected genes
}
\description{
Filter gene expression table
}

