% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{find_modules}
\alias{find_modules}
\title{Co-expression modules definition}
\usage{
find_modules(exprs, cor_method = c("pearson", "spearman"), min_ngen = 30,
  merge_similar = T, diss_thresh = 0.8, verbose = F)
}
\arguments{
\item{exprs}{Gene expression \code{data.frame}.}

\item{cor_method}{A character string indicating which correlation coefficient
is to be computed.}

\item{verbose}{Logical. If \code{TRUE}, reports analysis steps.}
}
\value{
just god knows
}
\description{
Defines co-expression modules
}
\examples{
coex_mods <- find_modules(exprs=expression.df, cor_method='pearson')

}

