% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\docType{methods}
\name{fit_data}
\alias{fit_data}
\alias{fit_data,CEMiTool-method}
\title{Retrieve scale-free model fit data}
\usage{
fit_data(cem)

\S4method{fit_data}{CEMiTool}(cem)
}
\arguments{
\item{cem}{Object of class \code{CEMiTool}}
}
\value{
Object of class \code{data.frame}
}
\description{
Retrieve scale-free model fit data
}
\examples{
# Get example CEMiTool object
data(cem)
# Get modules and beta data
cem <- find_modules(cem)
# Get fit data
fit_data(cem)

}
