% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\docType{methods}
\name{generate_report}
\alias{generate_report}
\alias{generate_report,CEMiTool-method}
\title{CEMiTool report}
\usage{
generate_report(cem_obj, ...)

\S4method{generate_report}{CEMiTool}(cem_obj, title = "CEMiTool Report",
  directory = "./reports")
}
\arguments{
\item{cem_obj}{Object of class \code{CEMiTool}.}

\item{title}{Report title.}

\item{directory}{Directory name for results.}
}
\description{
Creates report for CEMiTool results
}
\examples{

}

