% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-qq-line.R, R/stat-qq.R
\name{geom_qq_line}
\alias{geom_qq_line}
\alias{stat_qq_line}
\alias{geom_qq}
\alias{stat_qq}
\title{Geom qq}
\usage{
geom_qq_line(mapping = NULL, data = NULL, geom = "path",
  position = "identity", ..., distribution = stats::qnorm,
  dparams = list(), line.p = c(0.25, 0.75), fullrange = FALSE,
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE)

stat_qq_line(mapping = NULL, data = NULL, geom = "path",
  position = "identity", ..., distribution = stats::qnorm,
  dparams = list(), line.p = c(0.25, 0.75), fullrange = FALSE,
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE)

geom_qq(mapping = NULL, data = NULL, geom = "point",
  position = "identity", ..., distribution = stats::qnorm,
  dparams = list(), na.rm = FALSE, show.legend = NA, inherit.aes = TRUE)

stat_qq(mapping = NULL, data = NULL, geom = "point",
  position = "identity", ..., distribution = stats::qnorm,
  dparams = list(), na.rm = FALSE, show.legend = NA, inherit.aes = TRUE)
}
\arguments{
\item{mapping}{Parameter to ggplot}

\item{data}{Parameter to ggplot}

\item{geom}{Parameter to ggplot}

\item{position}{Parameter to ggplot}

\item{...}{Parameter to ggplot}

\item{distribution}{Distribution function to use, if x not specified}

\item{dparams}{Additional parameters passed on to `distribution`
function.}

\item{line.p}{Vector of quantiles to use when fitting the Q-Q line, defaults
defaults to `c(.25, .75)`.}

\item{fullrange}{Should the q-q line span the full range of the plot, or just
the data}

\item{na.rm}{Parameter to ggplot}

\item{show.legend}{Parameter to ggplot}

\item{inherit.aes}{Parameter to ggplot}
}
\value{
ggplot
}
\description{
Geom qq

`geom_qq` and `stat_qq` produce quantile-quantile plots. `geom_qq_line` and
`stat_qq_line` compute the slope and intercept of the line connecting the
points at specified quartiles of the theoretical and sample distributions. These
functions have been taken from the development version of ggplot2 and will be
removed as soon as they are implemented in the release version.
}
\examples{
\dontrun{
df <- data.frame(y = rt(200, df = 5))
p <- ggplot(df, aes(sample = y))
p + stat_qq() + stat_qq_line()

# Use fitdistr from MASS to estimate distribution params
params <- as.list(MASS::fitdistr(df$y, "t")$estimate)
ggplot(df, aes(sample = y)) +
  stat_qq(distribution = qt, dparams = params["df"]) +
  stat_qq_line(distribution = qt, dparams = params["df"])

# Using to explore the distribution of a variable
ggplot(mtcars, aes(sample = mpg)) +
  stat_qq() +
  stat_qq_line()
ggplot(mtcars, aes(sample = mpg, colour = factor(cyl))) +
  stat_qq() +
  stat_qq_line()
}
}
