% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\docType{methods}
\name{get_hubs}
\alias{get_hubs}
\alias{get_hubs,CEMiTool-method}
\title{Get Hubs}
\usage{
get_hubs(cem, ...)

\S4method{get_hubs}{CEMiTool}(cem, n = 5)
}
\arguments{
\item{cem}{Object of class \code{CEMiTool}.}

\item{...}{Optional parameters.}

\item{n}{Number of genes to return in each module (default: 5).}
}
\value{
A \code{list} containing hub genes.
}
\description{
Returns \code{n} genes in each module with high connectivity.
}
\examples{
\dontrun{
# Get example expression data
data(expr)
# Initialize CEMiTool object with expression
cem <- new('CEMiTool', expression=expr)
# Filter data
cem <- filter_expr(cem)
# Define network modules
cem <- find_modules(cem)
# Get module hubs
hubs <- get_hubs(cem, n=10)
}
}
