% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{graphplot}
\alias{graphplot}
\title{Expression profile visualization}
\usage{
graphplot(test)
}
\arguments{
\item{test}{}
}
\value{
None
}
\description{
Creates a line plot of each gene inside the module through the samples
}
\examples{
graphplot(test)

}

