% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{gseaplot}
\alias{gseaplot}
\title{GSEA visualization}
\usage{
gseaplot(test)
}
\arguments{
\item{test}{}
}
\value{
None
}
\description{
Creates a heatmap with the results of gene set enrichment analysis (GSEA) of co-expression modules
}
\examples{
gseaplot(test)

}

