% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{include_interaction}
\alias{include_interaction}
\title{Includes interaction information to CEMiTool object}
\usage{
include_interaction(cem_obj, ...)
}
\arguments{
\item{cem_obj}{Object of class \code{CEMiTool}.}

\item{int_df}{a data.frame or matrix containing two columns}
}
\value{
Object of class \code{CEMiTool}
}
\description{
Includes interaction information to CEMiTool object
}
\examples{

}

