% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\docType{methods}
\name{include_interactions}
\alias{include_interactions}
\alias{include_interactions,CEMiTool-method}
\title{Includes interaction information to CEMiTool object}
\usage{
include_interactions(cem, ...)

\S4method{include_interactions}{CEMiTool}(cem, int_df, directed = FALSE, ...)
}
\arguments{
\item{cem}{Object of class \code{CEMiTool}.}

\item{...}{parameters for igraph::graph_from_data_frame}

\item{int_df}{a data.frame or matrix containing two columns}

\item{directed}{Is the graph directed ? Default is FALSE}
}
\value{
Object of class \code{CEMiTool}
}
\description{
Includes interaction information to CEMiTool object
}
\examples{

}
