% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{mod_activity}
\alias{mod_activity}
\title{# Determine module activity based on limma}
\usage{
mod_activity(analyses, comp_group, subject_col)
}
\arguments{
\item{analyses}{List of \code{CEMiTool} objects}

\item{comp_group}{Character string indicating the group to be compared against}

\item{subject_col}{Optional character string indicating a column in the \code{CEMiTool} objects'
sample annotation slot object containing subject information}
}
\value{
A \code{data.frame} containing the results
}
\description{
# Determine module activity based on limma
}
\keyword{internal}
