% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cemitool.R
\docType{methods}
\name{mod_colors}
\alias{mod_colors}
\alias{mod_colors,CEMiTool-method}
\alias{mod_colors<-}
\alias{mod_colors<-,CEMiTool-method}
\title{Retrieve and set mod_colors attribute}
\usage{
mod_colors(cem_obj)

\S4method{mod_colors}{CEMiTool}(cem_obj)

mod_colors(cem_obj) <- value

\S4method{mod_colors}{CEMiTool}(cem_obj) <- value
}
\arguments{
\item{cem_obj}{Object of class \code{CEMiTool}}

\item{value}{a character vector containing colors for each module
the names should match with module names}
}
\value{

}
\description{
Retrieve and set mod_colors attribute
}

