% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{mod_compare}
\alias{mod_compare}
\title{Overlap statistics}
\usage{
mod_compare(analyses, p_thresh = 1, fdr_thresh = 1, jac_thresh = 0)
}
\arguments{
\item{analyses}{List of \code{CEMiTool} objects}

\item{p_thresh}{p-value threshold}

\item{fdr_thresh}{Threshold for FDR}

\item{jac_thresh}{Threshold for Jaccard}
}
\value{
A \code{data.frame} containing the results
}
\description{
Overlap statistics
}
\keyword{internal}
