% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{mod_gsea}
\alias{mod_gsea}
\title{Module Gene Set Enrichment Analysis}
\usage{
mod_gsea(test)
}
\arguments{
\item{test}{}
}
\value{
None
}
\description{
Perfoms gene set enrichment analysis for each co-expression module found
}
\examples{
mod_gsea(test)

}

