% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{mod_info}
\alias{mod_info}
\title{Compute node information and study information}
\usage{
mod_info(analyses, df_output, gsea_metric = "nes")
}
\arguments{
\item{analyses}{List of \code{CEMiTool} objects}

\item{df_output}{Output of function \code{mod_compare}}

\item{gsea_metric}{GSEA metric to be used. One of "nes",
"es" or "padj". Default: "nes".}
}
\value{
A \code{data.frame} containing the results
}
\description{
Compute node information and study information
}
\keyword{internal}
