% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{mod_ora}
\alias{mod_ora}
\title{Module Overrepresentation Analysis}
\usage{
mod_ora(test)
}
\arguments{
\item{test}{}
}
\value{
None
}
\description{
Perfoms overrepresentation analysis for each co-expression module found
}
\examples{
mod_ora(test)

}

