% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{mod_summary}
\alias{mod_summary}
\title{Co-expression module summarization}
\usage{
mod_summary(exprs, gene_module, method = c("mean", "eigengene"),
  verbose = F)
}
\arguments{
\item{exprs}{Gene expression \code{data.frame}.}

\item{gene_module}{Two column \code{data.frame}. First column with
gene identifiers and second column with module information.}

\item{verbose}{Logical. If \code{TRUE}, reports analysis steps.}
}
\value{
A data.frame with summarized values.
}
\description{
Summarizes modules using some statistics.
}
\examples{
mod_summary <- mod_summary(exprs=exprs, gene_module=gene_module)

}

