% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cemitool.R
\docType{methods}
\name{module_genes}
\alias{module_genes}
\alias{module_genes,CEMiTool-method}
\title{Get the module genes in a CEMiTool object}
\usage{
module_genes(cem)

\S4method{module_genes}{CEMiTool}(cem)
}
\arguments{
\item{cem}{Object of class \code{CEMiTool}}
}
\value{
Object of class \code{data.frame} containing genes and their
respective module
}
\description{
Get the module genes in a CEMiTool object
}
\examples{
# Get example CEMiTool object
cem <- CEMiTool::cem
# Get the module genes
module_genes(cem)

}
