% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cemitool.R
\docType{methods}
\name{module_names}
\alias{module_names}
\alias{module_names,CEMiTool-method}
\title{Get module names in a CEMiTool object}
\usage{
module_names(cem, include_NC = TRUE)

\S4method{module_names}{CEMiTool}(cem, include_NC = TRUE)
}
\arguments{
\item{cem}{Object of class \code{CEMiTool}}

\item{include_NC}{Logical. Whether or not to include "Not.Correlated" 
module. Defaults to \code{TRUE}.}
}
\value{
Module names
}
\description{
Get module names in a CEMiTool object
}
\examples{
# Get example CEMiTool object
cem <- CEMiTool::cem
# Get module names
module_names(cem)

}
