% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\docType{methods}
\name{module_report}
\alias{module_report}
\alias{module_report,CEMiTool-method}
\title{Module analysis report}
\usage{
module_report(cem_obj, ...)

\S4method{module_report}{CEMiTool}(cem_obj, directory)
}
\arguments{
\item{cem_obj}{Object of class \code{CEMiTool}.}

\item{directory}{Directory name for results.}
}
\description{
Creates report for modules analysis results
}
\examples{

}

