% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cemitool.R
\name{module_to_gmt}
\alias{module_to_gmt}
\title{Transform module genes list to a gmt file}
\usage{
module_to_gmt(cem, directory = "./Tables")
}
\arguments{
\item{cem}{}
}
\value{
A .gmt file containing module genes in each row
}
\description{
Transform module genes list to a gmt file
}
\keyword{internal}
