% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cemitool.R
\name{new_cem}
\alias{new_cem}
\title{Create a CEMiTool object}
\usage{
new_cem(expr = data.frame(), sample_annot = data.frame())
}
\arguments{
\item{expr}{Object of class \code{data.frame} with gene
expression data}

\item{sample_annot}{Object of \code{data.frame} containing the sample annotation.
It should have at least two columns containing group Class
and the Sample Name that should match with samples in
expression file}
}
\value{
Object of class \code{CEMiTool}
}
\description{
Create a CEMiTool object
}
\examples{
# Create new CEMiTool object
cem <- new_cem()
# Create new CEMiTool object with expression and sample_annotation data
data(expr)
data(sample_annot)
cem <- new_cem(expr, sample_annot)
# Equivalent to a call to new()
cem2 <- new("CEMiTool", expression=expr, sample_annotation=sample_annot)
identical(cem, cem2)
}
