% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cemitool.R
\name{number_modules.cemitool}
\alias{number_modules.cemitool}
\title{Get the number of modules on a cemitool object}
\usage{
\method{number_modules}{cemitool}(x)
}
\arguments{
\item{x}{cemitool object}
}
\value{
number of modules
}
\description{
Get the number of modules on a cemitool object
}

