% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\docType{methods}
\name{ora_data}
\alias{ora_data}
\alias{ora_data,CEMiTool-method}
\title{Retrieve over representation analysis (ORA) results}
\usage{
ora_data(cem)

\S4method{ora_data}{CEMiTool}(cem)
}
\arguments{
\item{cem}{Object of class \code{CEMiTool}}
}
\value{
Object of class \code{data.frame} with ORA data
}
\description{
Retrieve over representation analysis (ORA) results
}
\examples{
# Get example CEMiTool object
cem <- CEMiTool::cem
# Read gmt file
gmt <- read_gmt(system.file('extdata', 'pathways.gmt',
                   package='CEMiTool'))
# Run module overrepresentation analysis
cem <- mod_ora(cem, gmt)
# Check results
head(ora_data(cem))
}
