% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\docType{methods}
\name{ora_report}
\alias{ora_report}
\alias{ora_report,CEMiTool-method}
\title{Over-representation analysis report}
\usage{
ora_report(cem_obj, ...)

\S4method{ora_report}{CEMiTool}(cem_obj, directory, columns = c("ID", "Count",
  "GeneRatio", "BgRatio", "p.adjust"))
}
\arguments{
\item{cem_obj}{Object of class \code{CEMiTool}.}

\item{directory}{Directory name for results.}

\item{columns}{Character vector with column names to be shown in report.}
}
\description{
Creates report for modules over-representation analysis results
}
\examples{

}

