% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{oraplot}
\alias{oraplot}
\title{ORA visualization}
\usage{
oraplot(test)
}
\arguments{
\item{test}{}
}
\value{
None
}
\description{
Creates a bar plot with the results of overenrichment analysis of co-expression modules
}
\examples{
oraplot(test)

}

