% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{outer_join_merge}
\alias{outer_join_merge}
\title{Calculate full outer join for two ffdf objects}
\usage{
outer_join_merge(x, y)
}
\arguments{
\item{x}{An \code{ffdf} object}

\item{y}{An \code{ffdf} object}
}
\value{
A \code{ffdf} object containing the join result
}
\description{
Calculate full outer join for two ffdf objects
}
\keyword{internal}
