% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{overlap_community}
\alias{overlap_community}
\title{Generates communities from edgelist}
\usage{
overlap_community(mod_intersection_df, presence_as_weights = FALSE,
  smallest_community = 15, method = c("cluster_fast_greedy",
  "cluster_edge_betweenness", "cluster_label_prop",
  "cluster_leading_eigen", "cluster_louvain", "cluster_optimal",
  "cluster_spinglass", "cluster_walktrap"))
}
\arguments{
\item{mod_intersection_df}{Module intersection dataframe obtained from cemoverlap function}

\item{presence_as_weights}{Logical. Should sums of node pair occurence be considered edge weights?}

\item{smallest_community}{Minimal number of genes in community (default:15)}

\item{method}{Character string denoting an \code{igraph} package clustering function.
One of 'cluster_fast_greedy',  'cluster_edge_betweenness', 'cluster_fast_greedy',
 'cluster_label_prop', 'cluster_leading_eigen', 'cluster_louvain', 'cluster_optimal',
 'cluster_spinglass' or 'cluster_walktrap'. Default: 'cluster_fast_greedy'}
}
\value{
A list containing the genes present in each community detected
}
\description{
Returns communities from edgelist created by cemoverlap function.
}
\details{
Function takes edgelist as inputs and generates communities using functions
   provided in igraph package (default:'cluster_fast_greedy')
}
\examples{
\dontrun{
# Run the cemitool function twice on expr dataset. Each time, one sample will be removed
data(expr0)
set.seed(10)
dset1 <- expr0[,-sample(1:ncol(expr0), 1)]
dset2 <- expr0[,-sample(1:ncol(expr0), 1)]
cem1 <- cemitool(dset1, plot=FALSE)
cem2 <- cemitool(dset2, plot=FALSE)
cem_overlap_df <- cem_overlap(list(cem1, cem2))
comm_overlap_df <- overlap_community(cem_overlap_df)
}
}
