% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{plot_comembership}
\alias{plot_comembership}
\title{Plot module edge co-membership}
\usage{
plot_comembership(cem_overlap_df)
}
\arguments{
\item{cem_overlap_df}{Output of function \code{cem_overlap}}
}
\value{
A plot containing the number of edges present in each fraction of studies
}
\description{
Plot module edge co-membership
}
\examples{
# Run the cemitool function twice on expr dataset. Each time, one sample will be removed
data(expr0)
set.seed(10)
dset1 <- expr0[,-sample(1:ncol(expr0), 1)]
dset2 <- expr0[,-sample(1:ncol(expr0), 1)]
cem1 <- cemitool(dset1) 
cem2 <- cemitool(dset2) 
cem_overlap_df <- cem_overlap(list(cem1, cem2))
plot_comembership(cem_overlap_df)
}
