% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{plot_consensus}
\alias{plot_consensus}
\title{Plot graph of consensus modules}
\usage{
plot_consensus(cem_overlap_df, comm_overlap_df, study_num,
  num_sd_cut = 2)
}
\arguments{
\item{cem_overlap_df}{Output \code{data.frame} from function
\code{cem_overlap}}

\item{comm_overlap_df}{Output \code{data.frame} from function
\code{overlap_community}}

\item{study_num}{Minimum number of studies an edge must be present in for it
to be included}

\item{num_sd_cut}{Number of standard deviations an edge's mean must be above in order
to be included in the final plot (Default: 2)}
}
\value{
A plot containing the consensus modules of the studies
}
\description{
Plot graph of consensus modules
}
\examples{
\dontrun{
# Run the cemitool function twice on expr dataset. Each time, one sample will be removed
data(expr0)
data(sample_annot)
set.seed(10)
dset1 <- expr0[,-sample(1:ncol(expr0), 1)]
dset2 <- expr0[,-sample(1:ncol(expr0), 1)]
cem1 <- cemitool(dset1, sample_annot, plot=FALSE)
cem2 <- cemitool(dset2, sample_annot, plot=FALSE)
cem_overlap_df <- cem_overlap(list(cem1, cem2))
comm_overlap_df <- overlap_community(cem_overlap_df)
plot_consensus(cem_overlap_df, comm_overlap_df, study_num=2)
}
}
