% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_graph}
\alias{plot_graph}
\title{Expression profile visualization}
\usage{
plot_graph(edgelist)
}
\arguments{
\item{edgelist}{Two column \code{data.frame} with genes interaction data}
}
\value{
List with one network plot per module in gene_module
}
\description{
Creates a line plot of each gene inside the module through the samples
}
\examples{
plot_graph(edgelist)

}

