% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_interactions}
\alias{plot_interactions}
\title{Network visualization}
\usage{
plot_interactions(cem_obj, ...)
}
\arguments{
\item{cem_obj}{Object of class \code{CEMiTool}.}

\item{...}{Optional parameters.}

\item{n}{number of nodes to label}
}
\value{
Object of class \code{CEMiTool} with profile plots
}
\description{
Creates a graph based on interactions provided
}
\examples{
plot_profile(cem_obj)

}

