% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_ora}
\alias{plot_ora}
\title{ORA visualization}
\usage{
plot_ora(ora_res, n = 10, ...)
}
\arguments{
\item{ora_res}{a data.frame from ora function}

\item{n}{number of modules to show}

\item{...}{paramaters to plot_ora_single}
}
\value{
a list with ggplot2 object and the number of significant gene sets
}
\description{
Creates a bar plot with the results of overenrichment analysis of co-expression modules
}
\examples{
plot_ora(test)

}

