% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_profile}
\alias{plot_profile}
\title{Expression profile visualization}
\usage{
plot_profile(exprs, gene_module, annot = NULL, sample_col = NULL,
  class_col = NULL, order = TRUE, filename = "profile.pdf")
}
\arguments{
\item{exprs}{Expression data frame or list of expression data frames of each 
module.}

\item{gene_module}{Two column \code{data.frame}. First column with
gene identifiers and second column with module information}

\item{annot}{Optional. Two column \code{data.frame}. Fist column with
sample identifiers and second column with group/class information}

\item{sample_col}{character string with the name of samples column in the
annotation \code{data.frame}. If annot is given, this parameters
required.}

\item{class_col}{character string with the name of classes column in the
annotation \code{data.frame}. If annot is given, this parameters
required.}
}
\value{
List with one profile plot per module in gene_module
}
\description{
Creates a line plot of each gene inside the module through the samples
}
\examples{
plot_profile(exprs, gene_module)

}

