% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\docType{methods}
\name{plot_sample_tree}
\alias{plot_sample_tree}
\alias{plot_sample_tree,CEMiTool-method}
\title{Sample clustering}
\usage{
plot_sample_tree(obj, ...)

\S4method{plot_sample_tree}{CEMiTool}(obj, annot = NULL, col_vector = NULL,
  sample_name_column = NULL, class_column = NULL, filtered = FALSE)
}
\arguments{
\item{obj}{Object of class \code{CEMiTool} or \code{data.frame}.}

\item{...}{Optional parameters.}

\item{annot}{Annotation file. For CEMiTool objects, must be left NULL. See Details.}

\item{col_vector}{A vector of columns to use for visualizing the clustering. See Details.}

\item{sample_name_column}{A string specifying the column int the sample_annotation file
to be used as sample identification. For CEMiTool objects, this will be the string specified 
in the sample_name_column slot.}

\item{class_column}{A string specifying the column in the sample annotation file which indicates
sample groups. For CEMiTool objects, this will be the string specified in the class_column slot.}

\item{filtered}{Logical. Whether or not to use filtered data for CEMiTool objects. Default: FALSE.}
}
\value{
Object of class \code{CEMiTool} with dendrogram or a plot object.
}
\description{
Creates a dendrogram showing the similarities between samples in the expression data.
Both CEMiTool objects or data.frames can be used; for CEMiTool objects, the sample_annotation
slot data is used for creating the heatmap below the dendrogram. For a data.frame, annotation
can be provided via the annot argument.
}
\examples{
# Get example CEMiTool object
data(cem)
# Plot sample dendrogram
plot_sample_tree(cem)

}
