% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{plot_similarity}
\alias{plot_similarity}
\title{Plot study module similarity}
\usage{
plot_similarity(mod_stats, weight_col = "logfdr")
}
\arguments{
\item{mod_stats}{List output from function \code{stat_overlap_mods}}

\item{weight_col}{Character string denoting the weighting column for module similarity.
One of "Jaccard", "Fisherp", "fdr", "logp" or "logfdr". Default: "logfdr".}
}
\value{
A plot showing the similarity between study modules
}
\description{
Plot study module similarity
}
\examples{
\dontrun{
# Run the cemitool function five times on expr0 dataset. Each time, 10 samples will be removed.
data(expr0)
data(sample_annot)
set.seed(10)
dset1 <- expr0[,-sample(1:ncol(expr0), 10)]
set.seed(11)
dset2 <- expr0[,-sample(1:ncol(expr0), 10)]
set.seed(12)
dset3 <- expr0[,-sample(1:ncol(expr0), 10)]
set.seed(13)
dset4 <- expr0[,-sample(1:ncol(expr0), 10)]
set.seed(14)
dset5 <- expr0[,-sample(1:ncol(expr0), 10)]

cem1 <- cemitool(dset1, sample_annot, plot=FALSE)
cem2 <- cemitool(dset2, sample_annot, plot=FALSE)
cem3 <- cemitool(dset3, sample_annot, plot=FALSE)
cem4 <- cemitool(dset4, sample_annot, plot=FALSE)
cem5 <- cemitool(dset5, sample_annot, plot=FALSE)
mod_stats <- stat_overlap_mods(list(cem1, cem2, cem3, cem4, cem5), comp_group="g0")
plot_similarity(mod_stats)
}
}
