% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cemitool.R
\name{print.cemitool}
\alias{print.cemitool}
\title{Print a cemitool object}
\usage{
\method{print}{cemitool}(x)
}
\arguments{
\item{x}{cemitool object}
}
\value{
nothing
}
\description{
Print a cemitool object
}

