% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{profileplot}
\alias{profileplot}
\title{Expression profile visualization}
\usage{
profileplot(exprs)
}
\arguments{
\item{exprs}{expression data frame or list of expression data frames of each module}
}
\value{
None
}
\description{
Creates a line plot of each gene inside the module through the samples
}
\examples{
profileplot(exprs)

}

