% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{read_gmt}
\alias{read_gmt}
\title{Reads a GMT file}
\usage{
read_gmt(fname)
}
\arguments{
\item{fname}{file name of the gmt}

\item{fname}{GMT file name}
}
\value{
a list containing genes and description of each pathway
}
\description{
Reads a GMT file
}

