% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{refine_modules}
\alias{refine_modules}
\title{Co-expression module refinement}
\usage{
refine_modules(exprs, gene_module, nperm = 1000, use_lpc = F, verbose = F)
}
\arguments{
\item{exprs}{gene expression \code{data.frame}}

\item{gene_module}{two column \code{data.frame}. First column with
gene identifiers and second column with module information}

\item{nperm}{numeric. Number of permutations for edge significance test}

\item{use_lpc}{logical. If TRUE uses Local Partial Correlation method for
module refinement}

\item{verbose}{logical. Report analysis steps}
}
\value{
A list of with the following components:

gene_module: \code{data.frame} with gene identifier and module information
edge_list: \code{data.frame} with the genes which consist of each edge in 
           a module.
}
\description{
Refines modules by excluding non-significant edges.
}
\examples{
refined_mods <- refine_modules(exprs=expression.df, gene_module=coex)

}

