% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cemitool.R
\docType{methods}
\name{sample_annotation}
\alias{sample_annotation}
\alias{sample_annotation,CEMiTool-method}
\alias{sample_annotation<-}
\alias{sample_annotation<-,CEMiTool-method}
\title{Retrive or set the sample_annotation attribute}
\usage{
sample_annotation(cem_obj)

\S4method{sample_annotation}{CEMiTool}(cem_obj)

sample_annotation(cem_obj) <- value

\S4method{sample_annotation}{CEMiTool}(cem_obj) <- value
}
\arguments{
\item{cem_obj}{Object of class \code{CEMiTool}}

\item{value}{a data.frame containing the sample annotation, 
should have at least two columns containing the Class 
and the Sample Name that should match with samples in 
expression}
}
\description{
Retrive or set the sample_annotation attribute
}

