% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sample_annotation_data}
\alias{sample_annotation_data}
\title{Sample Annotation Dataset}
\description{
This dataset, together with \code{expr}
can be used as input for CEMiTool functions
}
\keyword{datasets}
