% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cemitool.R
\docType{methods}
\name{write_files}
\alias{write_files}
\alias{write_files,CEMiTool-method}
\title{Save the CEMiTool object in files}
\usage{
write_files(cem, directory)

\S4method{write_files}{CEMiTool}(cem, directory)
}
\arguments{
\item{cem}{Object of class \code{CEMiTool}}

\item{directory}{a directory}
}
\value{

}
\description{
Save the CEMiTool object in files
}

