% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\docType{methods}
\name{save_plot}
\alias{save_plot}
\alias{save_plot,CEMiTool-method}
\title{Save CEMiTool object plots}
\usage{
save_plot(cem, value, directory, force)

\S4method{save_plot}{CEMiTool}(cem, value = c("all", "profile", "gsea", "ora",
  "interaction", "beta_r2", "mean_k"), directory = "./Plots", force = FALSE)
}
\arguments{
\item{cem}{Object of class \code{CEMiTool}.}

\item{value}{A character string containing the name of the plot to be saved.}

\item{directory}{Directory into which the files will be saved.}

\item{force}{If the directory exists, execution will not stop.
One of "profile", "gsea", "ora", "interaction", "beta_r2", "mean_k" or "all".}
}
\value{
A pdf file with the desired plots
}
\description{
Save plots into the directory specified by the \code{directory} argument.
}
\examples{
# Get example CEMiTool object
data(cem)
# Plot beta x R squared graph
cem <- plot_beta_r2(cem)
# Save plot
\dontrun{save_plot(cem, value="beta_r2", directory="./Plots")}
}
