% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cemitool.R
\docType{methods}
\name{show,CEMiTool-method}
\alias{show,CEMiTool-method}
\title{Print a cemitool object}
\usage{
\S4method{show}{CEMiTool}(object)
}
\arguments{
\item{object}{Object of class CEMiTool}
}
\description{
Print a cemitool object
}
