% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{stat_overlap_mods}
\alias{stat_overlap_mods}
\title{Calculate module overlap statistics}
\usage{
stat_overlap_mods(analyses, comp_group, subject_col = NULL, ...)
}
\arguments{
\item{analyses}{List of \code{CEMiTool} objects}

\item{comp_group}{Character string indicating the group to be compared against}

\item{subject_col}{Optional character string indicating a column in the \code{CEMiTool}
objects' sample annotation slot object containing subject information}

\item{...}{Additional parameters}
}
\value{
A list containing overlap statistics, node and study information and module activity
}
\description{
Calculate module overlap statistics
}
\examples{
\dontrun{
# Run the cemitool function twice on expr dataset. Each time, one sample will be removed
data(expr0)
set.seed(10)
dset1 <- expr0[,-sample(1:ncol(expr0), 1)]
dset2 <- expr0[,-sample(1:ncol(expr0), 1)]
cem1 <- cemitool(dset1, sample_annot, plot=FALSE)
cem2 <- cemitool(dset2, sample_annot, plot=FALSE)
mod_stats <- stat_overlap_mods(analyses=list(cem1, cem2), comp_group="g0")
}
}
