% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{vst}
\alias{vst}
\title{Perform variance stabilizing transformation on expression file.}
\usage{
vst(expr)
}
\arguments{
\item{expr}{expression file containing genes in the rows and samples in the columns}
}
\value{
A data.frame containing the results.
}
\description{
Perform variance stabilizing transformation on expression file.
}
\keyword{internal}
