% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cemitool.R
\docType{methods}
\name{write_files}
\alias{write_files}
\alias{write_files,CEMiTool-method}
\title{Save the CEMiTool object in files}
\usage{
write_files(cem, ...)

\S4method{write_files}{CEMiTool}(cem, directory = "./Tables",
  force = FALSE)
}
\arguments{
\item{cem}{Object of class \code{CEMiTool}}

\item{...}{Optional parameters}

\item{directory}{a directory}

\item{force}{if the directory exists the execution will not stop}
}
\value{
A directory containing CEMiTool results in files.
}
\description{
Save the CEMiTool object in files
}
\examples{
# Get example CEMiTool object
data(cem)
# Save CEMiTool results in files
write_files(cem, directory=".", force=TRUE)

}
