% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cem}
\alias{cem}
\title{CEMiTool Object}
\format{An object of class \code{CEMiTool}}
\usage{
data(cem)
}
\description{
This object can be used as input for CEMiTool functions. Data used are from 
\code{expr} and \code{sample_annot}.
}
\examples{
# Get example CEMiTool object
data(cem)
cem
}
\keyword{data}
