% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cemitool.R
\docType{class}
\name{CEMiTool-class}
\alias{CEMiTool-class}
\title{An S4 class to represent the CEMiTool analysis.}
\description{
An S4 class to represent the CEMiTool analysis.
}
\section{Slots}{

\describe{
\item{\code{expression}}{Gene expression \code{data.frame}.}

\item{\code{sample_annotation}}{Sample annotation \code{data.frame}.}

\item{\code{module}}{Genes in modules information \code{data.frame}.}

\item{\code{enrichment}}{\code{list} with modules enrichment results for sample classes.}

\item{\code{ora}}{Over-representation analysis results \code{data.frame}.}

\item{\code{profile_plot}}{list of ggplot graphs with gene expression profile per module.}

\item{\code{enrichment_plot}}{ggplot graph for enrichment analysis results}

\item{\code{barplot_ora}}{list of ggplot graphs with over-representation analysis resultsper module}
}}

