% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cemitool.R
\name{cemitool}
\alias{cemitool}
\title{Full gene co-expression analysis}
\usage{
cemitool(exprs, annot, gmt, cor_method = c("pearson", "spearman"),
  cor_pval = 0.05, split_mods = F, ora_pval = 0.05, nperm = 1000,
  min_ngen = 30, diss_thresh = 0.8, plot = F, verbose = F)
}
\arguments{
\item{exprs}{gene expression \code{"data.frame"}}

\item{annot}{sample annotation \code{"data.frame"}}

\item{gmt}{gene set file in .gmt format}

\item{cor_method}{a character string indicating which correlation coefficient is to be computed}

\item{cor_pval}{p-value for gene-gene correlation}

\item{ora_pval}{p-value for overrepresentation analysis}

\item{nperm}{number of permutations}

\item{min_ngen}{minimal number of genes per submodule}

\item{diss_thresh}{module merging correlation threshold for eigengene similarity}

\item{verbose}{logical. Report analysis steps}

\item{ppi}{protein-protein interaction file}
}
\value{
just god knows
}
\description{
Defines co-expression modules and functionally characterizes
each one of them.
}
\examples{
cemitool(exprs=expression.df)

}

