% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cemitool.R
\name{cemitool}
\alias{cemitool}
\title{Full gene co-expression analysis}
\usage{
cemitool(expr, annot, gmt, interactions, filter = TRUE, filter_pval = 0.1,
  n_genes, cor_method = c("pearson", "spearman"),
  sample_name_column = "SampleName", class_column = "Class",
  merge_similar = TRUE, split_modules = FALSE, ora_pval = 0.05,
  min_ngen = 30, diss_thresh = 0.8, plot = FALSE, directed = FALSE,
  verbose = FALSE)
}
\arguments{
\item{expr}{Gene expression \code{data.frame}.}

\item{annot}{Sample annotation \code{data.frame}.}

\item{gmt}{a list from function prepare_gmt containing the gene sets.}

\item{interactions}{a data.frame containing two columns with gene names.}

\item{filter}{logical. If TRUE, will filter expression data.}

\item{filter_pval}{P-value threshold for filtering.}

\item{n_genes}{Number of genes left after filtering.}

\item{cor_method}{A character string indicating which correlation coefficient is
to be computed. One of \code{"pearson"} or \code{"spearman"}.
Default \code{"pearson"}.}

\item{sample_name_column}{A character string indicating the sample column 
name of the annotation table.}

\item{class_column}{A character string indicating the class column name of the 
annotation table.}

\item{merge_similar}{Logical. If \code{TRUE}, merge similar modules.}

\item{split_modules}{Logical. If \code{TRUE}, splits modules by correlation sign.}

\item{ora_pval}{P-value for overrepresentation analysis. Default \code{0.05}.}

\item{min_ngen}{Minimal number of genes per submodule. Default \code{30}.}

\item{diss_thresh}{Module merging correlation threshold for eigengene similarity.
Default \code{0.8}.}

\item{plot}{Logical. If \code{TRUE}, plots all figures.}

\item{directed}{Logical. If \code{TRUE}, the igraph objects in interactions slot will be directed.}

\item{verbose}{Logical. If \code{TRUE}, reports analysis steps.}
}
\value{
a cemitool object
}
\description{
Defines co-expression modules and functionally characterizes
each one of them.
}
\examples{
cemitool(expr=expr)

}

