% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{cemoverlap}
\alias{cemoverlap}
\title{Integrates CEMiTool analyses}
\usage{
cemoverlap(analyses)
}
\arguments{
\item{analyses}{List of objects of class \code{CEMiTool}}
}
\description{
Returns the occurence of edges between different analyses
}
\examples{
studies <- list('GSE12345'=cem1, 'GSE54321'=cem2)
overlaps <- cemoverlap(studies)

}

