% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{cemoverlap}
\alias{cemoverlap}
\title{Integrates CEMiTool analyses}
\usage{
cemoverlap(..., analyses_list = NULL, fraction = 1)
}
\arguments{
\item{Analyses}{List of objects of class \code{CEMiTool}}

\item{Fraction}{of objects an edge pair must be present to be selected}
}
\description{
Returns the occurrence of edges between different analyses
}
