% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cemitool.R
\docType{methods}
\name{expr_data}
\alias{expr_data}
\alias{expr_data,CEMiTool-method}
\alias{expr_data<-}
\alias{expr_data<-,CEMiTool-method}
\title{Retrieve and set expression attribute}
\usage{
expr_data(cem, ...)

\S4method{expr_data}{CEMiTool}(cem, filtered = TRUE)

expr_data(cem) <- value

\S4method{expr_data}{CEMiTool}(cem) <- value
}
\arguments{
\item{cem}{Object of class \code{CEMiTool}}

\item{filtered}{logical. If TRUE retrieves filtered expression data}

\item{value}{Object of class \code{data.frame} with gene
expression data}
}
\value{
Object of class \code{data.frame} with gene expression data
}
\description{
Retrieve and set expression attribute
}

