% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\docType{methods}
\name{filter_expr}
\alias{filter_expr}
\alias{filter_expr,CEMiTool-method}
\title{Filter gene expression table}
\usage{
filter_expr(cem, ...)

\S4method{filter_expr}{CEMiTool}(cem, pval = 0.1, n_genes, pct = 0.75,
  dtype = c("microarray", "rnaseq"))
}
\arguments{
\item{cem}{Object of class \code{CEMiTool}.}

\item{...}{Optional parameters.}

\item{pval}{P-value cutoff for gene selection.}

\item{n_genes}{Number of genes to be selected.}

\item{dtype}{Type of experiment. One of \code{'microarray'} or
\code{'rnaseq'}. Default is \code{'microarray'}.}
}
\value{
Object of class \code{CEMiTool} with selected genes
}
\description{
Filter gene expression table
}
\examples{

}

