% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\docType{methods}
\name{filter_expr}
\alias{filter_expr}
\alias{filter_expr,CEMiTool-method}
\title{Filter gene expression table}
\usage{
filter_expr(cem, ...)

\S4method{filter_expr}{CEMiTool}(cem, pval = 0.1, n_genes, pct = 0.75,
  apply_vst = FALSE)
}
\arguments{
\item{cem}{Object of class \code{CEMiTool}.}

\item{...}{Optional parameters.}

\item{pval}{P-value cutoff for gene selection.}

\item{n_genes}{Number of genes to be selected.}

\item{pct}{Percentage of most expressed genes to keep (before variance filter).}

\item{apply_vst}{Logical. If TRUE, will apply variance stabilizing transform before filtering data.}
}
\value{
Object of class \code{CEMiTool} with selected genes
}
\description{
Filter gene expression table
}
\examples{
# Get example expression data
data(expr)
# Initialize CEMiTool object with expression
cem <- new("CEMiTool", expression=expr)
# Filter genes
cem1 <- filter_expr(cem)
# Check selected genes
head(expr_data(cem1))
# Filter genes and apply variance stabilizing transformation
cem2 <- filter_expr(cem, apply_vst=TRUE)
# Check results
head(expr_data(cem2))

}
