% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{find_modules}
\alias{find_modules}
\title{Co-expression modules definition}
\usage{
find_modules(exprs, cor_method = c("pearson", "spearman"), min_ngen = 30,
  merge_similar = T, diss_thresh = 0.8, verbose = F)
}
\arguments{
\item{exprs}{Gene expression \code{data.frame}.}

\item{cor_method}{A character string indicating which correlation coefficient
is to be computed.}

\item{min_ngen}{Minimal number of genes per submodule. Default \code{30}.}

\item{merge_similar}{Logical. If \code{TRUE}, merge similar modules.}

\item{diss_thresh}{Module merging correlation threshold for eigengene similarity.
Default \code{0.8}.}

\item{verbose}{Logical. If \code{TRUE}, reports analysis steps.}
}
\value{
just god knows
}
\description{
Defines co-expression modules
}
\examples{
gene_module <- find_modules(exprs=exprs, cor_method='pearson')

}

