% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\docType{methods}
\name{get_hubs}
\alias{get_hubs}
\alias{get_hubs,CEMiTool-method}
\title{Get Hubs}
\usage{
get_hubs(cem_obj, ...)

\S4method{get_hubs}{CEMiTool}(cem_obj, n = 5)
}
\arguments{
\item{cem_obj}{Object of class \code{CEMiTool}.}

\item{n}{Number of genes to return in each module (default: 5).}
}
\value{
A \code{list} containing hub genes.
}
\description{
Returns \code{n} genes in each module with high connectivity.
}
\examples{

}

