% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cemitool.R
\docType{methods}
\name{mod_colors}
\alias{mod_colors}
\alias{mod_colors,CEMiTool-method}
\alias{mod_colors<-}
\alias{mod_colors<-,CEMiTool-method}
\title{Retrieve and set mod_colors attribute}
\usage{
mod_colors(cem)

\S4method{mod_colors}{CEMiTool}(cem)

mod_colors(cem) <- value

\S4method{mod_colors}{CEMiTool}(cem) <- value
}
\arguments{
\item{cem}{Object of class \code{CEMiTool}}

\item{value}{a character vector containing colors for each module
the names should match with module names}
}
\description{
Retrieve and set mod_colors attribute
}

