% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cemitool.R
\docType{methods}
\name{mod_colors}
\alias{mod_colors}
\alias{mod_colors,CEMiTool-method}
\alias{mod_colors<-}
\alias{mod_colors<-,CEMiTool,character-method}
\title{Retrieve and set mod_colors attribute}
\usage{
mod_colors(cem)

\S4method{mod_colors}{CEMiTool}(cem)

mod_colors(cem) <- value

\S4method{mod_colors}{CEMiTool,character}(cem) <- value
}
\arguments{
\item{cem}{Object of class \code{CEMiTool}}

\item{value}{a character vector containing colors for each module.
Names should match with module names}
}
\value{
A vector with color names.
}
\description{
Retrieve and set mod_colors attribute
}
\examples{
# Get example CEMiTool object
data(cem)
# See module colors
mod_colors(cem)
}
