% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{mod_gsea}
\alias{mod_gsea}
\title{Module Gene Set Enrichment Analysis}
\usage{
mod_gsea(exprs, gene_module, annot, sample_col = 1, class_col = 2,
  verbose = F)
}
\arguments{
\item{exprs}{Gene expression \code{data.frame}.}

\item{gene_module}{Two column \code{data.frame}. First column with
gene identifiers and second column with module information.}
}
\value{
GSEA results.
}
\description{
Perfoms gene set enrichment analysis for each co-expression module found.
}
\examples{
mod_gsea(test)

}

