% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{mod_gsea}
\alias{mod_gsea}
\title{Module Gene Set Enrichment Analysis}
\usage{
mod_gsea(exprs, gene_module, annot, sample_col = 1, class_col = 2,
  verbose = F)
}
\arguments{
\item{exprs}{}

\item{gene_module}{}
}
\value{

}
\description{
Perfoms gene set enrichment analysis for each co-expression module found
}
\examples{
mod_gsea(test)

}

