% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\docType{methods}
\name{mod_gsea}
\alias{mod_gsea}
\alias{mod_gsea,CEMiTool-method}
\title{Module Gene Set Enrichment Analysis}
\usage{
mod_gsea(cem, ...)

\S4method{mod_gsea}{CEMiTool}(cem, rank_method = "mean", verbose = FALSE)
}
\arguments{
\item{cem}{Object of class \code{CEMiTool}.}

\item{...}{Optional parameters.}

\item{rank_method}{Character string indicating how to rank genes. Either "mean" 
(the default) or "median".}

\item{verbose}{logical. Report analysis steps.}
}
\value{
GSEA results.
}
\description{
Perfoms Gene Set Enrichment Analysis (GSEA) for each co-expression module found.
}
\examples{
# Get example CEMiTool object
data(cem)
# Look at example annotation file
sample_annotation(cem)
# Run GSEA on network modules
cem <- mod_gsea(cem)
# Check results
gsea_data(cem)

}
\seealso{
\code{\link{plot_gsea}}
}
