% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\docType{methods}
\name{mod_gsea}
\alias{mod_gsea}
\alias{mod_gsea,CEMiTool-method}
\title{Module Gene Set Enrichment Analysis}
\usage{
mod_gsea(cem_obj, ...)

\S4method{mod_gsea}{CEMiTool}(cem_obj, verbose = F)
}
\arguments{
\item{cem_obj}{Object of class \code{CEMiTool}.}
}
\value{
GSEA results.
}
\description{
Perfoms gene set enrichment analysis for each co-expression module found.
}
\examples{
mod_gsea(cem_obj)

}

