% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\docType{methods}
\name{mod_gsea}
\alias{mod_gsea}
\alias{mod_gsea,CEMiTool-method}
\title{Module Gene Set Enrichment Analysis}
\usage{
mod_gsea(cem, ...)

\S4method{mod_gsea}{CEMiTool}(cem, verbose = FALSE)
}
\arguments{
\item{cem}{Object of class \code{CEMiTool}.}

\item{...}{Optional parameters.}

\item{verbose}{logical. Report analysis steps.}
}
\value{
GSEA results.
}
\description{
Perfoms gene set enrichment analysis for each co-expression module found.
}
\examples{
mod_gsea(cem)

}
