% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\docType{methods}
\name{mod_ora}
\alias{mod_ora}
\alias{mod_ora,CEMiTool-method}
\title{Module Overrepresentation Analysis}
\usage{
mod_ora(cem_obj, ...)

\S4method{mod_ora}{CEMiTool}(cem_obj, gmt_in, verbose = FALSE)
}
\arguments{
\item{cem_obj}{Object of class \code{CEMiTool}.}

\item{gmt_in}{Output of CEMiTool::read_gmt function.}

\item{verbose}{logical. Report analysis steps.}
}
\value{
Object of class \code{CEMiTool}
}
\description{
Perfoms overrepresentation analysis for each co-expression module found.
}
\examples{
mod_ora(cem_obj, read_gmt(system.file('extdata','pathways.gmt',
        package='CEMiTool')))

}

