% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{mod_ora}
\alias{mod_ora}
\title{Module Overrepresentation Analysis}
\usage{
mod_ora(gene_module, gmt, verbose = FALSE)
}
\arguments{
\item{gene_module}{Two column \code{data.frame}. First column with
gene identifiers and second column with module information.}

\item{gmt}{GMT file name.}

\item{verbose}{logical. Report analysis steps.}
}
\value{
None
}
\description{
Perfoms overrepresentation analysis for each co-expression module found.
}
\examples{
mod_ora(gene_module, gmt)

}

