% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\docType{methods}
\name{mod_ora}
\alias{mod_ora}
\alias{mod_ora,CEMiTool-method}
\title{Module Overrepresentation Analysis}
\usage{
mod_ora(cem, ...)

\S4method{mod_ora}{CEMiTool}(cem, gmt_in, verbose = FALSE)
}
\arguments{
\item{cem}{Object of class \code{CEMiTool}.}

\item{...}{Optional parameters.}

\item{gmt_in}{gmt file as outputted by \code{read_gmt} function.}

\item{verbose}{logical. Report analysis steps.}
}
\value{
Object of class \code{CEMiTool}
}
\description{
Performs overrepresentation analysis for each co-expression module found.
}
\examples{
\dontrun{
# Get example expression data
data(expr)
# Initialize CEMiTool object with expression 
cem <- new("CEMiTool", expression=expr)
# Filter data
cem <- filter_expr(cem)
# Read gmt file
gmt <- read_gmt(system.file('extdata', 'pathways.gmt',
                   package='CEMiTool'))
# Find modules in the data
cem <- find_modules(cem)
# Run module overrepresentation analysis
cem <- mod_ora(cem, gmt)
# Check results
cem@ora
}

}
