% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cemitool.R
\docType{methods}
\name{nmodules}
\alias{nmodules}
\alias{nmodules,CEMiTool-method}
\title{Get the number of modules on a cemitool object}
\usage{
nmodules(cem)

\S4method{nmodules}{CEMiTool}(cem)
}
\arguments{
\item{cem}{Object of class \code{CEMiTool}}
}
\value{
number of modules
}
\description{
Get the number of modules on a cemitool object
}
\examples{
# Get example expression data
data(expr)
cem <- new("CEMiTool", expression=expr)
# Find modules in the data
cem <- find_modules(cem)
# Get the number of modules
nmodules(cem)

}
