% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cemitool.R
\docType{methods}
\name{nmodules}
\alias{nmodules}
\alias{nmodules,CEMiTool-method}
\title{Get the number of modules in a CEMiTool object}
\usage{
nmodules(cem)

\S4method{nmodules}{CEMiTool}(cem)
}
\arguments{
\item{cem}{Object of class \code{CEMiTool}}
}
\value{
number of modules
}
\description{
Get the number of modules in a CEMiTool object
}
\examples{
# Get example CEMiTool object
data(cem)
# Get the number of modules
nmodules(cem)

}
