% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_graph}
\alias{plot_graph}
\title{Expression profile visualization}
\usage{
plot_graph(test)
}
\arguments{
\item{test}{}
}
\value{
None
}
\description{
Creates a line plot of each gene inside the module through the samples
}
\examples{
plot_graph(test)

}

