% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_gsea}
\alias{plot_gsea}
\title{GSEA visualization}
\usage{
plot_gsea(test)
}
\arguments{
\item{test}{}
}
\value{
None
}
\description{
Creates a heatmap with the results of gene set enrichment analysis (GSEA) of co-expression modules
}
\examples{
plot_gsea(test)

}

