% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_gsea}
\alias{plot_gsea}
\title{GSEA visualization}
\usage{
plot_gsea(enrichment, pv_cut = 0.05)
}
\arguments{
\item{enrichment}{Output of mod_gsea function.}

\item{pv_cut}{P-value cut-off. Default \code{0.05}}
}
\value{
None
}
\description{
Creates a heatmap with the results of gene set enrichment analysis (GSEA) of co-expression modules
}
\examples{
plot_gsea(enrichment)

}

