% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\docType{methods}
\name{plot_mean_k}
\alias{plot_mean_k}
\alias{plot_mean_k,CEMiTool-method}
\title{Network mean connectivity}
\usage{
plot_mean_k(cem, ...)

\S4method{plot_mean_k}{CEMiTool}(cem, title = "Mean connectivity")
}
\arguments{
\item{cem}{Object of class \code{CEMiTool}.}
}
\value{
Object of class \code{CEMiTool} with connectivity plot
}
\description{
Creates a graph showing the mean connectivity of genes in the network
}
\examples{
plot_mean_k(cem)

}
