% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_ora}
\alias{plot_ora}
\title{ORA visualization}
\usage{
plot_ora(test)
}
\arguments{
\item{test}{}
}
\value{
None
}
\description{
Creates a bar plot with the results of overenrichment analysis of co-expression modules
}
\examples{
plot_ora(test)

}

