% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\docType{methods}
\name{plot_ora}
\alias{plot_ora}
\alias{plot_ora,CEMiTool-method}
\title{ORA visualization}
\usage{
plot_ora(cem, ...)

\S4method{plot_ora}{CEMiTool}(cem, n = 10, ...)
}
\arguments{
\item{cem}{Object of class \code{CEMiTool}.}

\item{...}{paramaters to plot_ora_single}

\item{n}{number of modules to show}
}
\value{
Object of class \code{CEMiTool} with ORA plots
}
\description{
Creates a bar plot with the results of module overrepresentation analysis
}
\examples{
# Get example expression data
data(expr)
# Read example gmt file
gmt <- read_gmt(system.file('extdata', 'pathways.gmt',
                   package='CEMiTool'))
# Initialize CEMiTool object with expression data
cem <- new("CEMiTool", expression=expr)
# Find modules in the data
cem <- find_modules(cem)
# Run overrepresentation analysis 
# Plot module gene expression profiles
cem <- plot_ora(cem)
# Check results
cem@barplot_ora

}

