% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\docType{methods}
\name{plot_ora}
\alias{plot_ora}
\alias{plot_ora,CEMiTool-method}
\title{ORA visualization}
\usage{
plot_ora(cem_obj, ...)

\S4method{plot_ora}{CEMiTool}(cem_obj, n = 10, ...)
}
\arguments{
\item{cem_obj}{Object of class \code{CEMiTool}.}

\item{...}{paramaters to plot_ora_single}

\item{n}{number of modules to show}
}
\value{
Object of class \code{CEMiTool} with ORA plots
}
\description{
Creates a bar plot with the results of overenrichment analysis of co-expression modules
}
\examples{
plot_ora(test)

}

