% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\docType{methods}
\name{plot_profile}
\alias{plot_profile}
\alias{plot_profile,CEMiTool-method}
\title{Expression profile visualization}
\usage{
plot_profile(cem, ...)

\S4method{plot_profile}{CEMiTool}(cem, order = TRUE)
}
\arguments{
\item{cem}{Object of class \code{CEMiTool}.}

\item{...}{Optional parameters.}

\item{order}{Logical. If TRUE, sorts samples by class (only if CEMiTool object contains sample_annotation data.}
}
\value{
Object of class \code{CEMiTool} with profile plots
}
\description{
Creates a plot with module gene expression profiles along samples
}
\examples{
# Get example CEMiTool object
cem <- CEMiTool::cem
# Plot module gene expression profiles
cem <- plot_profile(cem)
# Check resulting plot
show_plot(cem, "profile")

}
