% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\docType{methods}
\name{plot_profile}
\alias{plot_profile}
\alias{plot_profile,CEMiTool-method}
\title{Expression profile visualization}
\usage{
plot_profile(cem_obj, ...)

\S4method{plot_profile}{CEMiTool}(cem_obj, order = TRUE)
}
\arguments{
\item{cem_obj}{Object of class \code{CEMiTool}.}

\item{...}{Optional parameters.}
}
\value{
Object of class \code{CEMiTool} with profile plots
}
\description{
Creates a line plot of each gene inside the module through the samples
}
\examples{
plot_profile(cem_obj)

}

