% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\docType{methods}
\name{plot_profile}
\alias{plot_profile}
\alias{plot_profile,CEMiTool-method}
\title{Expression profile visualization}
\usage{
plot_profile(cem, ...)

\S4method{plot_profile}{CEMiTool}(cem, order = TRUE)
}
\arguments{
\item{cem}{Object of class \code{CEMiTool}.}

\item{...}{Optional parameters.}

\item{order}{Logical. If TRUE, sorts samples by class.}
}
\value{
Object of class \code{CEMiTool} with profile plots
}
\description{
Creates a plot with module gene expression profiles along samples
}
\examples{
\dontrun{
# Get example expression data
data(expr)
# Get example sample annotation data
data(sample_annot)
# Initialize CEMiTool object only with expression 
cem <- new("CEMiTool", expression=expr)
# Filter data
cem <- filter_expr(cem)
# Find modules in the data
cem <- find_modules(cem)
# Plot module gene expression profiles
cem <- plot_profile(cem)

# Initialize CEMiTool object with expression and sample annotation
cem <- new("CEMiTool", expression=expr, sample_annotation=sample_annot)
# Find modules in the data
cem <- find_modules(cem)
# Plot module gene expression profiles with group classes as defined in sample_annot
cem <- plot_profile(cem)
}

}
