% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{read_gmt}
\alias{read_gmt}
\title{Read a GMT file}
\usage{
read_gmt(fname)
}
\arguments{
\item{fname}{GMT file name.}
}
\value{
A list containing genes and description of each pathway
}
\description{
Read a GMT file
}
\examples{
# Read example gmt file
gmt_fname <- system.file("extdata", "pathways.gmt", package = "CEMiTool")
gmt_in <- read_gmt(gmt_fname)

}
