% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cemitool.R
\docType{methods}
\name{sample_annotation}
\alias{sample_annotation}
\alias{sample_annotation,CEMiTool-method}
\alias{sample_annotation<-}
\alias{sample_annotation<-,CEMiTool-method}
\title{Retrive or set the sample_annotation attribute}
\usage{
sample_annotation(cem)

\S4method{sample_annotation}{CEMiTool}(cem)

sample_annotation(cem) <- value

\S4method{sample_annotation}{CEMiTool}(cem) <- value
}
\arguments{
\item{cem}{Object of class \code{CEMiTool}}

\item{value}{A data.frame containing the sample annotation,
should have at least two columns containing the Class
and the Sample Name that should match with samples in
expression}
}
\description{
Retrive or set the sample_annotation attribute
}
\examples{
# Get example expression data
data(expr)
# Get example sample_annotation data
data(sample_annot)
# Initialize CEMiTool object with expression
cem <- new("CEMiTool", expression=sample_annot)
# Add sample annotation file to CEMiTool object
sample_annotation(cem) <- sample_annot
# Check annotation file
head(cem@sample_annotation)

}
