% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\docType{methods}
\name{split_modules}
\alias{split_modules}
\alias{split_modules,CEMiTool-method}
\title{Co-expression module refinement}
\usage{
split_modules(cem, ...)

\S4method{split_modules}{CEMiTool}(cem, min_ngen = 30, verbose = FALSE)
}
\arguments{
\item{cem}{Object of class \code{CEMiTool}.}

\item{...}{Optional parameters.}

\item{min_ngen}{Minimal number of genes per submodule. Default \code{30}.}

\item{verbose}{Logical. If \code{TRUE}, reports analysis steps.}
}
\value{
An object of class \code{CEMiTool}.
}
\description{
Refines modules by splitting them into submodules based on correlation sign.
}
\examples{
splitted_mods <- split_modules(cem)

}

